Move Metrics related code to a dedicated namespace / directory.
