+++ Test/CoreSDK.Test/Shared/Extensibility/Metrics/MetricTest.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace Microsoft.ApplicationInsights.Extensibility
+namespace Microsoft.ApplicationInsights.Extensibility.Metrics
 {
     using System;
     using System.Collections.Generic;
