+++ src/Core/Managed/Shared/Extensibility/Implementation/Metrics/SimpleMetricStatisticsAggregator.cs
@@ -1,4 +1,4 @@
-﻿namespace Microsoft.ApplicationInsights.Extensibility.Implementation
+﻿namespace Microsoft.ApplicationInsights.Extensibility.Implementation.Metrics
 {
     using System;
     using System.Threading;
