+++ src/Core/Managed/Shared/Extensibility/Metrics/MetricManager.cs
@@ -1,4 +1,4 @@
-﻿namespace Microsoft.ApplicationInsights.Extensibility
+﻿namespace Microsoft.ApplicationInsights.Extensibility.Metrics
 {
     using System;
     using System.Collections.Concurrent;
@@ -10,6 +10,7 @@
     using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.ApplicationInsights.Extensibility.Implementation;
     using Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing;
+    using Implementation.Metrics;
 
 #if CORE_PCL || NET45 || NET46
     using TaskEx = System.Threading.Tasks.Task;
